+++ package.json
@@ -21,6 +21,7 @@
     "nanocolors": "^0.2.13",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-elastic-carousel": "^0.11.5",
     "react-hook-form": "^7.17.0",
     "react-icons": "^4.3.1",
     "react-redux": "^7.2.5",
