+++ src/pages/CustomOrder/customorderstyles.css
@@ -1,5 +1,31 @@
+@import url('https://fonts.googleapis.com/css2?family=Zen+Kaku+Gothic+New:wght@400;500&display=swap');
+
+h1 {
+  font-family: 'Zen Kaku Gothic New', sans-serif;
+  font-weight: bold;
+}
+
+h2 {
+  font-family: 'Zen Kaku Gothic New', sans-serif;
+}
+
+h3 {
+  font-family: 'Zen Kaku Gothic New', sans-serif;
+}
+
+h4 {
+  font-family: 'Zen Kaku Gothic New', sans-serif;
+}
+
+h5 {
+  font-family: 'Zen Kaku Gothic New', sans-serif;
+}
+
+h6 {
+  font-family: 'Zen Kaku Gothic New', sans-serif;
+}
 .product_custom_container {
-  margin: 2% 20%;
+  margin: 2% 10%;
 }
 
 .image_preview {
@@ -10,10 +36,15 @@
   margin-left: auto;
   margin-right: auto;
   margin-top: 8%;
+  margin-bottom: 50px;
+}
+
+.form_grid {
+  margin-top: 10%;
 }
 
 .form_container {
-  padding: 50px 15vw;
+  padding: 50px 10vw;
 }
 
 .btn_add_form {
