--- src/pages/Home/CustomOrder/customorderstyles.css
@@ -1,22 +1,0 @@
-.product_custom_container {
-  margin: 2% 20%;
-}
-
-.image_preview {
-  height: 400px;
-  width: 400px;
-  max-width: 100%;
-  max-height: 100%;
-  margin-left: auto;
-  margin-right: auto;
-  margin-top: 8%;
-}
-
-.form_container {
-  padding: 50px 15vw;
-}
-
-.btn_add_form {
-  display: flex;
-  padding: 20% 50% 20% 0;
-}
