--- src/pages/Home/ProductDetail/productdetailstyles.css
@@ -1,17 +1,0 @@
-.product_detail_container {
-  margin: 7% 15%;
-}
-
-.image_container {
-  margin-top: 10%;
-  width: 400px;
-  height: 400px;
-  max-width: 100%;
-  max-height: 100%;
-}
-
-.product_container {
-  display: flex;
-  flex-direction: column;
-  margin: -44% 10% 20% 40%;
-}
