--- src/pages/Home/homestyles.css
@@ -1,11 +1,0 @@
-.input_search {
-  display: flex;
-  flex-direction: column;
-  align-items: left;
-  margin-bottom: 20px;
-}
-
-.products_container {
-  margin-left: 17%;
-  margin-right: 1%;
-}
