--- src/pages/Home/homestyles.js
@@ -1,13 +1,0 @@
-import { makeStyles } from '@material-ui/core/styles';
-
-export default makeStyles((theme) => ({
-  toolbar: theme.mixins.toolbar,
-  content: {
-    flexGrow: 1,
-    padding: theme.spacing(3),
-    marginLeft: '2.5%',
-  },
-  root: {
-    flexGrow: 1,
-  },
-}));
