+++ src/pages/index.js
@@ -1,7 +1,9 @@
 export { default as AdminDashboard } from './Admin/Admin';
 export { default as AdminProducts } from './AdminProducts/AdminProducts';
-export { default as Home } from './Home/Home';
+export { default as Products } from './Products/Products';
 export { default as ProductDetail } from './ProductDetail/ProductDetail';
 export { default as CustomOrder } from './CustomOrder/CustomOrder';
 export { default as UserTransactions } from './UserTransactions/UserTransactions';
 export { default as TransactionDetail } from './TransactionDetail/TransactionDetail';
+export { default as Sale } from './SaleProducts/SaleProducts';
+export { default as Home } from './Home/Home';
