+++ src/routes/AppRoute.js
@@ -10,7 +10,6 @@ import Register from '../pages/Auth/Register';
 import RegisterSuccess from '../pages/Auth/RegisterSuccess';
 import Verify from '../pages/Auth/Verify';
 import ChangePassword from '../pages/ChangePassword';
-import Home from '../pages/Home/Home';
 import Profile from '../pages/Profile';
 import {
   AdminDashboard,
@@ -19,6 +18,9 @@ import {
   CustomOrder,
   UserTransactions,
   TransactionDetail,
+  Products,
+  Sale,
+  Home,
 } from '../pages';
 import Transaction from '../pages/Transaction';
 import Cart from '../pages/Cart';
@@ -58,6 +60,8 @@ function AppRoute() {
           {/* PRODUCTS ROUTING */}
           <Route component={ProductDetail} path="/productdetail/:product_id" />
           <Route component={CustomOrder} path="/customorder" />
+          <Route component={Sale} path="/sale" />
+          <Route exact path="/products" component={Products} />
           <Route exact path="/transaction" component={Transaction} />
           <Route
             exact
